structure adjustments
