+++ app/Http/Controllers/LineController.php
@@ -13,6 +13,7 @@ use \LINE\LINEBot;
 use GuzzleHttp\Client;
 use App\User;
 use App\Leave;
+use App\Libraries\Helpers;
 
 class LineController extends Controller
 {
@@ -89,7 +90,7 @@ class LineController extends Controller
         return $json;
     }
 
-    public function lineEvent(Request $request)
+    public function lineEvent(Request $request, Helpers $helpers)
     {
         $lineApi = DB::table('credentials')->select('username as secret', 'password as access_token')->where('description', 'line_api')->first();
         $accessToken = $lineApi->access_token;
@@ -161,72 +162,72 @@ class LineController extends Controller
 
             // Approx Entry Time empty member list
             } elseif ($msgText == '進場統計') {
-                $members = User::where([['guild','無與倫比'],['approx_entry_time', '']])->get();
-                $memberCount = $members->count();
-                $memberList = '';
-                foreach ($members as $member) {
-                    $memberList .= "{$member->lineid}\n";
-                }
+                // $members = User::where([['guild','無與倫比'],['approx_entry_time', '']])->get();
+                // $memberCount = $members->count();
+                // $memberList = '';
+                // foreach ($members as $member) {
+                //     $memberList .= "{$member->lineid}\n";
+                // }
 
-                // Team Count
-                $tanhungTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '丹紅城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
-                $linmoTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '蓮慕城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
-                $choiloTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城'], ['approx_entry_time', '準時參加']])->count();
-                $taihoTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '大豪城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
-                $buffTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
+                // // Team Count
+                // $tanhungTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '丹紅城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
+                // $linmoTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '蓮慕城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
+                // $choiloTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城'], ['approx_entry_time', '準時參加']])->count();
+                // $taihoTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '大豪城'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
+                // $buffTeamCount = User::where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組'], ['approx_entry_time', '<>', ''], ['approx_entry_time', '準時參加']])->count();
 
-                // Team List
-                $choiloTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->count();
-                $choiloTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->get();
-                $choiloTeamLateList = '';
-                foreach ($choiloTeamLate as $choiloLateMember) {
-                    $choiloTeamLateList .= "{$choiloLateMember->lineid}|";
-                }
-                $choiloTeamLateList = rtrim($choiloTeamLateList, '|');
+                // // Team List
+                // $choiloTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->count();
+                // $choiloTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->get();
+                // $choiloTeamLateList = '';
+                // foreach ($choiloTeamLate as $choiloLateMember) {
+                //     $choiloTeamLateList .= "{$choiloLateMember->lineid}|";
+                // }
+                // $choiloTeamLateList = rtrim($choiloTeamLateList, '|');
 
-                $choiloTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->get();
-                $choiloTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->count();
-                $choiloTeamLeaveList = '';
-                foreach ($choiloTeamLeave as $choiloLeaveMember) {
-                    $choiloTeamLeaveList .= "{$choiloLeaveMember->lineid}|";
-                }
-                $choiloTeamLeaveList = rtrim($choiloTeamLeaveList, '|');
+                // $choiloTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->get();
+                // $choiloTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '塞羅城']])->count();
+                // $choiloTeamLeaveList = '';
+                // foreach ($choiloTeamLeave as $choiloLeaveMember) {
+                //     $choiloTeamLeaveList .= "{$choiloLeaveMember->lineid}|";
+                // }
+                // $choiloTeamLeaveList = rtrim($choiloTeamLeaveList, '|');
 
-                // Team Taiho
-                $taihoTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->count();
-                $taihoTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->get();
-                $taihoTeamLateList = '';
-                foreach ($taihoTeamLate as $taihoLateMember) {
-                    $taihoTeamLateList .= "{$taihoLateMember->lineid}|";
-                }
-                $taihoTeamLateList = rtrim($taihoTeamLateList, '|');
+                // // Team Taiho
+                // $taihoTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->count();
+                // $taihoTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->get();
+                // $taihoTeamLateList = '';
+                // foreach ($taihoTeamLate as $taihoLateMember) {
+                //     $taihoTeamLateList .= "{$taihoLateMember->lineid}|";
+                // }
+                // $taihoTeamLateList = rtrim($taihoTeamLateList, '|');
 
-                $taihoTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->get();
-                $taihoTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->count();
-                $taihoTeamLeaveList = '';
-                foreach ($taihoTeamLeave as $taihoLeaveMember) {
-                    $taihoTeamLeaveList .= "{$taihoLeaveMember->lineid}|";
-                }
-                $taihoTeamLeaveList = rtrim($taihoTeamLeaveList, '|');
+                // $taihoTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->get();
+                // $taihoTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '大豪城']])->count();
+                // $taihoTeamLeaveList = '';
+                // foreach ($taihoTeamLeave as $taihoLeaveMember) {
+                //     $taihoTeamLeaveList .= "{$taihoLeaveMember->lineid}|";
+                // }
+                // $taihoTeamLeaveList = rtrim($taihoTeamLeaveList, '|');
                 
-                // Team Buff
-                $buffTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->count();
-                $buffTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->get();
-                $buffTeamLateList = '';
-                foreach ($buffTeamLate as $buffLateMember) {
-                    $buffTeamLateList .= "{$buffLateMember->lineid}|";
-                }
-                $buffTeamLateList = rtrim($buffTeamLateList, '|');
+                // // Team Buff
+                // $buffTeamLateCount = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->count();
+                // $buffTeamLate = User::whereIn('approx_entry_time', array('晚到10分鐘', '晚到11~20分鐘', '晚到30分鐘以上'))->where([['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->get();
+                // $buffTeamLateList = '';
+                // foreach ($buffTeamLate as $buffLateMember) {
+                //     $buffTeamLateList .= "{$buffLateMember->lineid}|";
+                // }
+                // $buffTeamLateList = rtrim($buffTeamLateList, '|');
 
-                $buffTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->get();
-                $buffTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->count();
-                $buffTeamLeaveList = '';
-                foreach ($buffTeamLeave as $buffLeaveMember) {
-                    $buffTeamLeaveList .= "{$buffLeaveMember->lineid}|";
-                }
-                $buffTeamLeaveList = rtrim($buffTeamLeaveList, '|');
+                // $buffTeamLeave = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->get();
+                // $buffTeamLeaveCount = User::where([['approx_entry_time', '無法參加本次爭奪'], ['guild','無與倫比'],['guildwar_phase_1', '增益：鬼怪組']])->count();
+                // $buffTeamLeaveList = '';
+                // foreach ($buffTeamLeave as $buffLeaveMember) {
+                //     $buffTeamLeaveList .= "{$buffLeaveMember->lineid}|";
+                // }
+                // $buffTeamLeaveList = rtrim($buffTeamLeaveList, '|');
 
-                $response = $bot->replyText($replyToken, "未設定進場狀態({$memberCount}): \n{$memberList}\n各分組登記狀態: \n丹紅: {$tanhungTeamCount}\n蓮慕: {$linmoTeamCount}\n-------塞羅:({$choiloTeamCount})------\n晚到({$choiloTeamLateCount}):{$choiloTeamLateList}\n請假({$choiloTeamLeaveCount}):{$choiloTeamLeaveList}\n\n-------大豪:({$taihoTeamCount})------\n晚到({$taihoTeamLateCount}):{$taihoTeamLateList}\n請假({$taihoTeamLeaveCount}):{$taihoTeamLeaveList}\n\n-------鬼怪:({$buffTeamCount})------\n晚到({$buffTeamLateCount}):{$buffTeamLateList}\n請假({$buffTeamLeaveCount}):{$buffTeamLeaveList}");
+                $response = $bot->replyText($replyToken, $helpers->statistics());
 
             /*
              * ===============================
@@ -352,28 +353,11 @@ class LineController extends Controller
                 $response = $bot->replyText($replyToken, "指令列表: \n--- 查詢類 ---\n戰力排行\n爭奪券數\n戰力\n等級\n進場狀態\n進場統計\n\n--- 設定類 ---\n更新戰力:{數值}\n更新券數:{數值}\n準時\n晚10\n晚20\n晚30\n請假:{事由}\n");
 
             } else {
-                $randomText = [
-                    "小幫手壞掉了嗎？\n好像真的壞掉了耶{$this->emoji('0x10007B')}", 
-                    "這個指令小幫手暫時無法識別呢{$this->emoji('0x100091')}\n如果希望小幫手加入這項功能\n可以在門派群組提出建議哦{$this->emoji('0x10008D')}", 
-                    "指令列表裡面好像...\n沒有這個指令哦{$this->emoji('0x10008C')}", 
-                    "哈囉，今天小幫手休假哦\n{$this->emoji('0x100085')}小幫手怎麼會有休假?!\n好像說得也對吼{$this->emoji('0x10007C')}",
-                    "請確認指令後重試",
-                    "等一下，我打個電話給會長看看是不是把我的電源給踢掉了...",
-                    "我在看水行俠，等下看完回你哦{$this->emoji('0x100095')}"
-                ];
-                $response = $bot->replyText($replyToken, $randomText[mt_rand(0, count($randomText)-1)]);
+                // use random replies
+                $response = $bot->replyText($replyToken, $helpers->randomReply());
             }
         }
         return response('OK', 200);
-    }
-
-    public function emoji($code)
-    {
-        if ($code) {
-            $code = str_replace('0x', '', $code);
-            $bin = hex2bin(str_repeat('0', 8 - strlen($code)) . $code);
-            return mb_convert_encoding($bin, 'UTF-8', 'UTF-32BE');
-        }
     }
 }
